+++ extensions/amp-accordion/0.1/amp-accordion.js
@@ -19,7 +19,7 @@ import {isExperimentOn} from '../../../src/experiments';
 import {Layout} from '../../../src/layout';
 import {dev, user} from '../../../src/log';
 import {removeFragment} from '../../../src/url';
-import {map} from '../../../src/types';
+import {map} from '../../../src/utils/object';
 
 /** @const */
 const TAG = 'amp-accordion-session-state-optout';
